refactor: renamed package from npcs to patrons
