+++ src/main/java/spireCafe/interactables/patrons/purpletear/Clone.java
@@ -1,11 +1,11 @@
-package spireCafe.interactables.npcs.purpletear;
+package spireCafe.interactables.patrons.purpletear;
 
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import spireCafe.abstracts.AbstractSCCard;
 
 import static spireCafe.Anniv7Mod.makeID;
-import static spireCafe.interactables.npcs.purpletear.PurpleTearPatron.assetID;
+import static spireCafe.interactables.patrons.purpletear.PurpleTearPatron.assetID;
 
 public class Clone extends AbstractSCCard {
     public final static String ID = makeID(Clone.class.getSimpleName());
