+++ src/main/java/spireCafe/interactables/patrons/purpletear/Prescript.java
@@ -1,4 +1,4 @@
-package spireCafe.interactables.npcs.purpletear;
+package spireCafe.interactables.patrons.purpletear;
 
 import basemod.AutoAdd;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
@@ -6,7 +6,7 @@ import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import spireCafe.abstracts.AbstractSCCard;
 
 import static spireCafe.Anniv7Mod.makeID;
-import static spireCafe.interactables.npcs.purpletear.PurpleTearPatron.assetID;
+import static spireCafe.interactables.patrons.purpletear.PurpleTearPatron.assetID;
 
 @AutoAdd.Ignore
 public class Prescript extends AbstractSCCard {
